Player fix
Fix player jump and movement
