+++ Assets/Prefabs/Platforms/StaticPlatform.prefab
@@ -25,7 +25,7 @@ GameObject:
   - component: {fileID: 251644776575206014}
   m_Layer: 0
   m_Name: StaticPlatform
-  m_TagString: Untagged
+  m_TagString: Platform
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
