--- Assets/Sprites/Interfaz.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 4a4f52511103f47b2a6d6608d9c64a2f
-folderAsset: yes
-timeCreated: 1519492683
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
