--- Assets/Sprites/Personajes/Armado.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 09d8f5f278c51449e8af11e9066c7385
-folderAsset: yes
-timeCreated: 1519499629
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
