--- Assets/Sprites/Personajes/Saltador.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: a4e6967fcbdad45ddbf3a5c150c6dc28
-folderAsset: yes
-timeCreated: 1519499634
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
