--- Assets/Sprites/Personajes/Suicida.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 72ea96b518b904bb384bd5275d497e92
-folderAsset: yes
-timeCreated: 1519499462
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
