--- Assets/Sprites/Personajes/Volador.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 0eba9416391fc4158b634fe1ad0c5964
-folderAsset: yes
-timeCreated: 1519499663
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
